+++ font-end-admin/src/app/component/voucher/action-voucher/action-voucher.component.ts
@@ -5,6 +5,7 @@ import { DetailVoucherComponent } from '../detail-voucher/detail-voucher.compone
 import Swal from 'sweetalert2';
 import { VoucherComponent } from '../voucher.component';
 import { EditVoucherComponent } from '../edit-voucher/edit-voucher.component';
+import { ToastrService } from 'ngx-toastr';
 
 @Component({
   selector: 'app-action-voucher',
@@ -13,18 +14,24 @@ import { EditVoucherComponent } from '../edit-voucher/edit-voucher.component';
 })
 
 export class ActionVoucherComponent implements OnInit {
-  isMenuOpen: boolean = false;
+  // isMenuOpen: boolean = false;
   data: any;
   params: any;
+  role: 'ADMIN' | 'USER' | 'STAFF';
 
   constructor(
     private matdialog: MatDialog,
     private voucherService: VoucherService,
     private cdr: ChangeDetectorRef,
-    private voucherComponent: VoucherComponent
+    private voucherComponent: VoucherComponent,
+    private toastr: ToastrService,
   ) { }
 
-  ngOnInit(): void { }
+  ngOnInit(): void {
+    var userjson = localStorage.getItem("users");
+    var users = JSON.parse(userjson);
+    this.role = users.role;
+  }
 
   agInit(params: any) {
     this.params = params.data;
@@ -34,18 +41,37 @@ export class ActionVoucherComponent implements OnInit {
     return false;
   }
 
+  // openUpdate(id: number): void {
+  //   const dialogref = this.matdialog.open(EditVoucherComponent, {
+  //     width: '250vh',
+  //     height: '98vh',
+  //     data: { idVoucher: id },
+  //   });
+  //   dialogref.afterClosed().subscribe((result) => {
+  //     if (result === 'saveVoucher') {
+  //       this.voucherComponent.ngOnInit();
+  //       this.cdr.detectChanges();
+  //     }
+  //   });
+  // }
+
   openUpdate(id: number): void {
-    const dialogref = this.matdialog.open(EditVoucherComponent, {
-      width: '250vh',
-      height: '98vh',
-      data: { idVoucher: id },
-    });
-    dialogref.afterClosed().subscribe((result) => {
-      if (result === 'saveVoucher') {
-        this.voucherComponent.ngOnInit();
-        this.cdr.detectChanges();
-      }
-    });
+    if (this.role === 'ADMIN') {
+      const dialogRef = this.matdialog.open(EditVoucherComponent, {
+        width: '250vh',
+        height: '98vh',
+        data: { idVoucher: id },
+      });
+
+      dialogRef.afterClosed().subscribe((result) => {
+        if (result === 'saveVoucher') {
+          this.voucherComponent.ngOnInit();
+          this.cdr.detectChanges();
+        }
+      });
+    } else if (this.role === 'STAFF') {
+      this.toastr.error('Bạn không có quyền cập nhật', 'Lỗi');
+    }
   }
 
   clickDetail(id: number): void {
@@ -61,25 +87,51 @@ export class ActionVoucherComponent implements OnInit {
   }
 
   deleteVoucher(id?: number) {
-    Swal.fire({
-      title: 'Bạn có chắc muốn xóa',
-      text: 'Bạn sẽ không thể hoàn tác',
-      icon: 'warning',
-      showCancelButton: true,
-      confirmButtonColor: '#3085d6',
-      cancelButtonColor: '#d33',
-      confirmButtonText: 'Xóa',
-      cancelButtonText: 'Thoát',
-    }).then((response) => {
-      if (response.isConfirmed) {
-        this.voucherService.deleteVoucher(id).subscribe(() => {
-          this.voucherComponent.ngOnInit();
-          this.cdr.detectChanges();
-        });
-        Swal.fire('Xóa', 'Xóa thành công', 'success');
-      }
-    });
+    if (this.role === 'ADMIN') {
+      Swal.fire({
+        title: 'Bạn có chắc muốn xóa',
+        text: 'Bạn sẽ không thể hoàn tác',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Xóa',
+        cancelButtonText: 'Thoát',
+      }).then((response) => {
+        if (response.isConfirmed) {
+          this.voucherService.deleteVoucher(id).subscribe(() => {
+            this.voucherComponent.ngOnInit();
+            this.cdr.detectChanges();
+          });
+          Swal.fire('Xóa', 'Xóa thành công', 'success');
+        }
+      });
+    } else if (this.role === 'STAFF') {
+      this.toastr.error('Bạn không có quyền xóa', 'Lỗi'); // Show error message
+    }
   }
+
+
+  // deleteVoucher(id?: number) {
+  //   Swal.fire({
+  //     title: 'Bạn có chắc muốn xóa',
+  //     text: 'Bạn sẽ không thể hoàn tác',
+  //     icon: 'warning',
+  //     showCancelButton: true,
+  //     confirmButtonColor: '#3085d6',
+  //     cancelButtonColor: '#d33',
+  //     confirmButtonText: 'Xóa',
+  //     cancelButtonText: 'Thoát',
+  //   }).then((response) => {
+  //     if (response.isConfirmed) {
+  //       this.voucherService.deleteVoucher(id).subscribe(() => {
+  //         this.voucherComponent.ngOnInit();
+  //         this.cdr.detectChanges();
+  //       });
+  //       Swal.fire('Xóa', 'Xóa thành công', 'success');
+  //     }
+  //   });
+  // }
 
   updateIdel(id?: number) {
     this.voucherService.getDetailVoucher(id).subscribe((response: any[]) => {
